rework general methods to return completable future - replaced proto classes with correspond pojo classes
