+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.avenga</groupId>
     <artifactId>parent-steam-client</artifactId>
-    <version>1.0.3</version>
+    <version>1.1.0</version>
     <packaging>pom</packaging>
 
     <modules>
