+++ steam-client/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>com.avenga</groupId>
         <artifactId>parent-steam-client</artifactId>
-        <version>1.0.3</version>
+        <version>1.1.0</version>
     </parent>
 
     <artifactId>steam-client</artifactId>
