+++ steam-client/src/main/java/com/avenga/steamclient/steam/coordinator/callback/GCSessionCallbackHandler.java
@@ -9,19 +9,14 @@ import com.avenga.steamclient.steam.client.callback.AbstractCallbackHandler;
 
 public class GCSessionCallbackHandler extends AbstractCallbackHandler<GCPacketMessage> {
 
-    public static ClientGCProtobufMessage<CMsgClientWelcome.Builder> handle(SteamMessageCallback<GCPacketMessage> callback) {
-        GCPacketMessage gcPacketMessage = waitAndGetPacketMessage(callback, "GCSession");
-
-        return getMessage(gcPacketMessage);
-    }
-
-    public static ClientGCProtobufMessage<CMsgClientWelcome.Builder> handle(SteamMessageCallback<GCPacketMessage> callback, long timeout) throws CallbackTimeoutException {
+    public static CMsgClientWelcome handle(SteamMessageCallback<GCPacketMessage> callback, long timeout) throws CallbackTimeoutException {
         GCPacketMessage gcPacketMessage = waitAndGetPacketMessage(callback, timeout, "GCSession");
 
         return getMessage(gcPacketMessage);
     }
 
-    private static ClientGCProtobufMessage<CMsgClientWelcome.Builder> getMessage(GCPacketMessage gcPacketMessage) {
-        return new ClientGCProtobufMessage<>(CMsgClientWelcome.class, gcPacketMessage);
+    public static CMsgClientWelcome getMessage(GCPacketMessage gcPacketMessage) {
+        ClientGCProtobufMessage<CMsgClientWelcome.Builder> clientWelcome = new ClientGCProtobufMessage<>(CMsgClientWelcome.class, gcPacketMessage);
+        return clientWelcome.getBody().build();
     }
 }
