+++ steam-language-gen/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.avenga</groupId>
         <artifactId>parent-steam-client</artifactId>
-        <version>1.0.3</version>
+        <version>1.1.0</version>
     </parent>
 
     <artifactId>steam-language-gen</artifactId>
